#89: Refactor class locations
