+++ Frontend/ShoppingList.Frontend.Infrastructure/Connection/ICommandQueue.cs
@@ -1,5 +1,5 @@
-﻿using ProjectHermes.ShoppingList.Frontend.Models.Common;
-using ProjectHermes.ShoppingList.Frontend.Models.Shared.Requests;
+﻿using ProjectHermes.ShoppingList.Frontend.Models.Shared.Requests;
+using ProjectHermes.ShoppingList.Frontend.WebApp.Services.Error;
 using System.Threading.Tasks;
 
 namespace ProjectHermes.ShoppingList.Frontend.Infrastructure.Connection
