+++ Frontend/ShoppingList.Frontend.Infrastructure/Error/ICommandQueueErrorHandler.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace ProjectHermes.ShoppingList.Frontend.Models.Common
+namespace ProjectHermes.ShoppingList.Frontend.WebApp.Services.Error
 {
     public interface ICommandQueueErrorHandler
     {
