+++ Frontend/ShoppingList.Frontend.WebApp/Components/Common/DebugSwitch.razor
@@ -1,4 +1,4 @@
-﻿@using ProjectHermes.ShoppingList.Frontend.Models.Common;
+﻿@using ProjectHermes.ShoppingList.Frontend.WebApp.Services.Error;
 
 <div class="pr-2">
     <Switch Checked="@DebugHandler.IsDebug" CheckedChildren="Dbg" UnCheckedChildren="Dbg" OnChange="() => DebugHandler.ToggleDebugState()" />
