+++ Frontend/ShoppingList.Frontend.WebApp/Components/Common/ErrorStack.razor
@@ -1,4 +1,4 @@
-﻿@using ProjectHermes.ShoppingList.Frontend.Models.Index;
+﻿@using ProjectHermes.ShoppingList.Frontend.WebApp.Pages.Index.Models;
 
 <div>
     @foreach (var msg in ErrorReceiver.Stack)
