+++ Frontend/ShoppingList.Frontend.WebApp/Components/Index/Item.razor
@@ -1,6 +1,6 @@
 ﻿@using ProjectHermes.ShoppingList.Frontend.Models;
-@using ProjectHermes.ShoppingList.Frontend.Models.Index;
 @using ProjectHermes.ShoppingList.Frontend.Models.Shared.Requests;
+@using ProjectHermes.ShoppingList.Frontend.WebApp.Pages.Index.Models;
 
 @inject Microsoft.AspNetCore.Components.NavigationManager navigationManager;
 @inject ProjectHermes.ShoppingList.Frontend.Infrastructure.Connection.ICommandQueue commandQueue;
