+++ Frontend/ShoppingList.Frontend.WebApp/Components/Index/ItemSearchBar.razor
@@ -1,10 +1,10 @@
 ﻿@inject ProjectHermes.ShoppingList.Frontend.Infrastructure.Connection.IApiClient apiClient;
 @inject ProjectHermes.ShoppingList.Frontend.Infrastructure.Connection.ICommandQueue commandQueue;
-@inject Models.Common.Services.IShoppingListNotificationService notificationService;
+@inject ProjectHermes.ShoppingList.Frontend.WebApp.Services.Notification.IShoppingListNotificationService notificationService;
 @inject ProjectHermes.ShoppingList.Frontend.Models.Index.Services.ITemporaryItemCreationService temporaryItemCreationService;
 @using ProjectHermes.ShoppingList.Frontend.Models.Index.Search;
 @using ProjectHermes.ShoppingList.Frontend.Models.Shared.Requests;
-@using ProjectHermes.ShoppingList.Frontend.Models.Index;
+@using ProjectHermes.ShoppingList.Frontend.WebApp.Pages.Index.Models;
 @using ProjectHermes.ShoppingList.Frontend.Models.Shared;
 
 <div class="mb-2 d-flex">
