+++ Frontend/ShoppingList.Frontend.WebApp/Components/Index/ItemVisiblityButton.razor
@@ -1,4 +1,4 @@
-﻿@using ProjectHermes.ShoppingList.Frontend.Models.Index;
+﻿@using ProjectHermes.ShoppingList.Frontend.WebApp.Pages.Index.Models;
 
 <Button Class="mr-1 d-flex align-items-center"
         OnClick="@OnShowBasketItemsChangeButtonClickedAsync"
