+++ Frontend/ShoppingList.Frontend.WebApp/Components/Index/Section.razor
@@ -1,5 +1,5 @@
 ﻿@using ProjectHermes.ShoppingList.Frontend.Models;
-@using ProjectHermes.ShoppingList.Frontend.Models.Index;
+@using ProjectHermes.ShoppingList.Frontend.WebApp.Pages.Index.Models;
 
 <div>
     <div class="py-2 mb-1 pl-1 d-flex align-items-center cursor-pointer @GetBackgroundColor() @GetFontColor()" @onclick="OnSectionClicked">
