+++ Frontend/ShoppingList.Frontend.WebApp/Components/Index/StoreSelect.razor
@@ -1,7 +1,9 @@
 ﻿@inject ProjectHermes.ShoppingList.Frontend.Infrastructure.Connection.IApiClient apiClient;
-@inject Models.Common.Services.IShoppingListNotificationService notificationService;
+@inject ProjectHermes.ShoppingList.Frontend.WebApp.Services.Notification.IShoppingListNotificationService notificationService;
+
 @using ProjectHermes.ShoppingList.Frontend.Models;
 @using ProjectHermes.ShoppingList.Frontend.Models.Index;
+@using ProjectHermes.ShoppingList.Frontend.WebApp.Pages.Index.Models;
 
 <Select Class="min-w-100 max-w-200 mr-3"
         DataSource="availableStores.Stores"
