+++ Frontend/ShoppingList.Frontend.WebApp/Components/Items/Editor/ItemCategorySelect.razor
@@ -1,9 +1,9 @@
 ﻿@using ProjectHermes.ShoppingList.Frontend.Models;
 @using ProjectHermes.ShoppingList.Frontend.Models.Items;
-@using ProjectHermes.ShoppingList.Frontend.Models.Common.Error;
+@using ProjectHermes.ShoppingList.Frontend.WebApp.Services.Error;
 
 @inject ProjectHermes.ShoppingList.Frontend.Infrastructure.Connection.IApiClient apiClient;
-@inject ProjectHermes.ShoppingList.Frontend.Models.Common.Services.IShoppingListNotificationService notificationService;
+@inject ProjectHermes.ShoppingList.Frontend.WebApp.Services.Notification.IShoppingListNotificationService notificationService;
 
 <div>
     <div class="font-weight-bold">Category</div>
