+++ Frontend/ShoppingList.Frontend.WebApp/Components/Items/Filter/FilterResult.razor
@@ -1,5 +1,5 @@
 ﻿@using ProjectHermes.ShoppingList.Frontend.Models.Items;
-@using ProjectHermes.ShoppingList.Frontend.Models.Common.Error;
+@using ProjectHermes.ShoppingList.Frontend.WebApp.Services.Error;
 
 @inject ProjectHermes.ShoppingList.Frontend.WebApp.Services.Items.IItemsPageLoadingService loadingService;
