+++ Frontend/ShoppingList.Frontend.WebApp/Components/Items/Filter/ItemFilter.razor
@@ -1,6 +1,6 @@
 ﻿@using ProjectHermes.ShoppingList.Frontend.Models;
 @using ProjectHermes.ShoppingList.Frontend.Models.Items;
-@using ProjectHermes.ShoppingList.Frontend.Models.Common.Error;
+@using ProjectHermes.ShoppingList.Frontend.WebApp.Services.Error;
 
 @inject ProjectHermes.ShoppingList.Frontend.WebApp.Services.Items.IItemsPageLoadingService loadingService;
