+++ Frontend/ShoppingList.Frontend.WebApp/Components/Stores/StoreEditor.razor
@@ -1,6 +1,6 @@
 ﻿@using ProjectHermes.ShoppingList.Frontend.WebApp.Pages.Stores.Models;
 @using ProjectHermes.ShoppingList.Frontend.Models.Shared.Requests;
-@using ProjectHermes.ShoppingList.Frontend.Models.Common.Error;
+@using ProjectHermes.ShoppingList.Frontend.WebApp.Services.Error;
 
 @inject ProjectHermes.ShoppingList.Frontend.WebApp.Pages.Stores.Services.IStoresPageCommunicationService communicationService;
