+++ Frontend/ShoppingList.Frontend.WebApp/Pages/Items/Models/ItemsErrorHandler.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.AspNetCore.Components;
-using ProjectHermes.ShoppingList.Frontend.Models.Common.Error;
+using ProjectHermes.ShoppingList.Frontend.WebApp.Services.Error;
 using System;
 using System.Threading.Tasks;
 
-namespace ProjectHermes.ShoppingList.Frontend.Models.Items
+namespace ProjectHermes.ShoppingList.Frontend.WebApp.Pages.Items.Models
 {
     public class ItemsErrorHandler : IAsyncRetryFragmentCreator
     {
