+++ Frontend/ShoppingList.Frontend.WebApp/Pages/Items/Models/ItemsState.cs
