+++ Frontend/ShoppingList.Frontend.WebApp/Services/Error/IRetryFragmentCreator.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Components;
 using System;
 
-namespace ProjectHermes.ShoppingList.Frontend.Models.Common.Error
+namespace ProjectHermes.ShoppingList.Frontend.WebApp.Services.Error
 {
     public interface IRetryFragmentCreator
     {
