+++ Frontend/ShoppingList.Frontend.WebApp/Services/Notification/ShoppingListNotificationService.cs
@@ -1,8 +1,7 @@
 ﻿using AntDesign;
 using Microsoft.AspNetCore.Components;
-using ProjectHermes.ShoppingList.Frontend.Models.Common.Services;
 
-namespace ProjectHermes.ShoppingList.Frontend.WebApp.Services
+namespace ProjectHermes.ShoppingList.Frontend.WebApp.Services.Notification
 {
     public class ShoppingListNotificationService : IShoppingListNotificationService
     {
